+++ src/power_grid_model/__init__.py
@@ -4,9 +4,9 @@
 
 """Power Grid Model"""
 
-from power_grid_model.core.dataset_definitions import ComponentType, DatasetType
-from power_grid_model.core.power_grid_meta import initialize_array, power_grid_meta_data
-from power_grid_model.core.power_grid_model import PowerGridModel
+from power_grid_model._core.dataset_definitions import ComponentType, DatasetType
+from power_grid_model._core.power_grid_meta import initialize_array, power_grid_meta_data
+from power_grid_model._core.power_grid_model import PowerGridModel
 from power_grid_model.enum import (
     Branch3Side,
     BranchSide,
