+++ src/power_grid_model/_core/buffer_handling.py
@@ -12,6 +12,10 @@ from typing import cast
 
 import numpy as np
 
+from power_grid_model._core.error_handling import VALIDATOR_MSG
+from power_grid_model._core.index_integer import IdxC, IdxNp
+from power_grid_model._core.power_grid_core import IdxPtr, VoidPtr
+from power_grid_model._core.power_grid_meta import ComponentMetaData
 from power_grid_model._utils import (
     _extract_data_from_component_data,
     _extract_indptr,
@@ -19,10 +23,6 @@ from power_grid_model._utils import (
     is_columnar,
     is_sparse,
 )
-from power_grid_model.core.error_handling import VALIDATOR_MSG
-from power_grid_model.core.index_integer import IdxC, IdxNp
-from power_grid_model.core.power_grid_core import IdxPtr, VoidPtr
-from power_grid_model.core.power_grid_meta import ComponentMetaData
 from power_grid_model.data_types import (
     AttributeType,
     ComponentData,
