+++ src/power_grid_model/_core/index_integer.py
