+++ src/power_grid_model/_core/power_grid_meta.py
@@ -12,7 +12,7 @@ from typing import Any
 
 import numpy as np
 
-from power_grid_model.core.dataset_definitions import (
+from power_grid_model._core.dataset_definitions import (
     ComponentTypeLike,
     ComponentTypeVar,
     DatasetType,
@@ -20,7 +20,7 @@ from power_grid_model.core.dataset_definitions import (
     _str_to_component_type,
     _str_to_datatype,
 )
-from power_grid_model.core.power_grid_core import AttributePtr, ComponentPtr, DatasetPtr, power_grid_core as pgc
+from power_grid_model._core.power_grid_core import AttributePtr, ComponentPtr, DatasetPtr, power_grid_core as pgc
 from power_grid_model.data_types import DenseBatchArray, SingleArray
