+++ src/power_grid_model/typing.py
@@ -6,7 +6,7 @@
 Type hints for PGM. This includes all miscellaneous type hints not under dataset or dataset_definitions categories
 """
 
-from power_grid_model.core.dataset_definitions import ComponentType, ComponentTypeVar
+from power_grid_model._core.dataset_definitions import ComponentType, ComponentTypeVar
 from power_grid_model.enum import ComponentAttributeFilterOptions
 
 _ComponentAttributeMappingDict = dict[ComponentType, set[str] | list[str] | None | ComponentAttributeFilterOptions]
