+++ tests/unit/test_data_handling.py
@@ -7,12 +7,12 @@ import warnings
 import numpy as np
 import pytest
 
+from power_grid_model._core.data_handling import create_output_data
+from power_grid_model._core.dataset_definitions import ComponentType as CT, DatasetType as DT
+from power_grid_model._core.power_grid_core import VoidPtr
+from power_grid_model._core.power_grid_dataset import CMutableDataset
+from power_grid_model._core.power_grid_meta import initialize_array
 from power_grid_model._utils import is_columnar
-from power_grid_model.core.data_handling import create_output_data
-from power_grid_model.core.dataset_definitions import ComponentType as CT, DatasetType as DT
-from power_grid_model.core.power_grid_core import VoidPtr
-from power_grid_model.core.power_grid_dataset import CMutableDataset
-from power_grid_model.core.power_grid_meta import initialize_array
 
 
 @pytest.mark.parametrize(
