+++ tests/unit/test_error_handling.py
@@ -7,7 +7,7 @@ from copy import copy
 import pytest
 
 from power_grid_model import PowerGridModel
-from power_grid_model.core.power_grid_meta import initialize_array
+from power_grid_model._core.power_grid_meta import initialize_array
 from power_grid_model.enum import CalculationMethod, LoadGenType, MeasuredTerminalType, TapChangingStrategy
 from power_grid_model.errors import (
     AutomaticTapCalculationError,
