+++ tests/unit/test_utils.py
@@ -9,7 +9,7 @@ import numpy as np
 import pytest
 
 from power_grid_model import LoadGenType, initialize_array
-from power_grid_model.core.power_grid_meta import power_grid_meta_data
+from power_grid_model._core.power_grid_meta import power_grid_meta_data
 from power_grid_model.data_types import Dataset
 from power_grid_model.utils import (
     export_json_data,
