+++ tests/unit/validation/test_validation_functions.py
@@ -9,7 +9,7 @@ import numpy as np
 import pytest
 
 from power_grid_model import CalculationType, LoadGenType, MeasuredTerminalType, initialize_array, power_grid_meta_data
-from power_grid_model.core.dataset_definitions import ComponentType, DatasetType
+from power_grid_model._core.dataset_definitions import ComponentType, DatasetType
 from power_grid_model.enum import Branch3Side, BranchSide, CalculationType, FaultType, TapChangingStrategy
 from power_grid_model.validation import assert_valid_input_data
 from power_grid_model.validation.errors import (
